+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.0.37
 
+- Add Right Element and Expandable element in Compact-card-list
 - Fix Avatar component prop heirarchy
 - Fix PeopleTabs spacings
 - Fix PeopleListItem component UI add middle element